Move Configuration class into S3 as that's where it'll be used
